+++ src/lib/api.ts
@@ -433,14 +433,6 @@ export type Objective = {
   minimize: boolean;
 };
 
-export type NautilusObjectiveData = {
-  name: string;
-  minimize: boolean;
-  value: number;
-  color: string;
-  id: string;
-}[];
-
 export type Variable = {
   name: string;
 };
