+++ src/lib/components/methods/nautilus/types.ts
@@ -12,3 +12,23 @@ export enum AppState {
   IDLE = "IDLE",
   WORKING = "WORKING",
 }
+
+export type NautilusObjectiveData = {
+  name: string;
+  minimize: boolean;
+  value: number;
+  color: string;
+  id: string;
+}[];
+
+export type Iteration = {
+  currentIterationPoint: number[];
+  distance: number[];
+  lowerBounds: number[];
+  upperBounds: number[];
+};
+
+export type NautilusRanks = {
+  name: string;
+  items: NautilusObjectiveData[];
+}[];
